payment_service/requirements.txt
flask
psycopg2-binary
stripe
python-dotenv
pytest
pytest-flask
